refactor: move fetchReceiptWithRetry and retry options to common module
